refactor: consolidate error types into shared package
- Add upstreamResponse property to UpstreamError class
- Include Claude-specific error code mapping
- Update serializeError with special UpstreamError handling
- Update all imports to use @claude-nexus/shared
- Remove duplicate proxy/src/types/errors.ts file
- Fix duplicate import issues
