+++ services/proxy/src/middleware/validation.ts
@@ -1,6 +1,5 @@
 import { Context, Next } from 'hono'
-import { ValidationError } from '../types/errors'
-import { validateClaudeRequest } from '@claude-nexus/shared'
+import { ValidationError, validateClaudeRequest } from '@claude-nexus/shared'
 import { getRequestLogger } from './logger'
 import { maskSensitiveData, truncateString } from '@claude-nexus/shared/utils/validation'
